MT#55283 scale delete-delay

+++ include/call.h
@@ -874,9 +874,9 @@ void dialogue_connect(struct call_monologue *, struct call_monologue *, sdp_ng_f
 bool monologue_transform(struct call_monologue *, sdp_ng_flags *, medias_q *);
 void monologue_destroy(struct call_monologue *ml);
 int call_delete_branch_by_id(const str *callid, const str *branch,
-	const str *fromtag, const str *totag, ng_command_ctx_t *, int delete_delay);
+	const str *fromtag, const str *totag, ng_command_ctx_t *, int64_t delete_delay);
 int call_delete_branch(call_t *, const str *branch,
-	const str *fromtag, const str *totag, ng_command_ctx_t *, int delete_delay);
+	const str *fromtag, const str *totag, ng_command_ctx_t *, int64_t delete_delay);
 void call_destroy(call_t *);
 struct call_media *call_media_new(call_t *call);
 void call_media_free(struct call_media **mdp);
